refactor and optimizated data space
