--- L101_022.h
@@ -1,4 +1,0 @@
-#ifndef KVA_H
-#define KVA_H
-void insertion_sort(int arr[], int n);
-#endif
